Update README.md
# Savvy-Solutions-Africa-Data-Cleaning-Solution
Data cleaning is a critical step in data preparation process which is aimed at ensuring the accuracy, consistency and the completeness of a data before commencing analysis.
The cleaning of data involves identifying and rectifying of errors ,anomalies and inconsistencies to further enhance the data quality and reliability. As a skilled data analyst, the combination of automated and manual techniques were used to refine and prepare the data for insightful analysis.

Key Process:
Data inspection and profiling (Initial Assessment of data and data profiling)
Handling of missing Data(Identification of missing values, imputation using the mean, median and mode depending on the nature of the data)
Removing Duplicates(Detection using algorithms, duplicates based on key attributes, consolidate and the elimination method)
Addressing of outliers
Data validation and quality checks.
In Conclusion, effective data cleaning was essentially done to produce reliable and actionable insights, by meticulously addressing data quality for subsequent analysis leading to more accurate and meaningful results.